requirements.txt
discord.py==2.3.2
aiohttp>=3.9.0
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
bilibili-cli>=0.3.0
webdriver-manager==4.0.1